+++ src/main/java/com/epam/esm/service/impl/GiftCertificateComparatorProvider.java
@@ -1,7 +1,6 @@
 package com.epam.esm.service.impl;
 
 import com.epam.esm.model.entity.GiftCertificate;
-import com.epam.esm.model.entity.SortType;
 
 import java.util.Comparator;
 import java.util.Optional;
@@ -33,5 +32,27 @@ class GiftCertificateComparatorProvider {
             optional = Optional.empty();
         }
         return optional;
+    }
+
+    private enum SortType {
+        NAME((c1, c2) -> c1.getName().compareTo(c2.getName())),
+        CREATE_DATE((c1, c2) -> c1.getCreateDate().compareTo(c2.getCreateDate())),
+        LAST_UPDATE_DATE((c1, c2) -> c1.getLastUpdateDate().compareTo(c2.getLastUpdateDate())),
+        PRICE((c1, c2) -> c1.getPrice() - c2.getPrice()),
+        DURATION((c1, c2) -> c1.getDuration() - c2.getDuration());
+
+        private Comparator<GiftCertificate> comparator;
+
+        SortType(Comparator<GiftCertificate> comparator) {
+            this.comparator = comparator;
+        }
+
+        public Comparator<GiftCertificate> getComparatorAsc() {
+            return comparator;
+        }
+
+        public Comparator<GiftCertificate> getComparatorDesc() {
+            return comparator.reversed();
+        }
     }
 }
